+++ src/containers/pages/InterventionPlan/NewPlan/General/tests/index.test.tsx
@@ -4,27 +4,41 @@ import { createBrowserHistory } from 'history';
 import React from 'react';
 import { Provider } from 'react-redux';
 import { Router } from 'react-router';
+import configureMockStore from 'redux-mock-store';
+import { defaultProps as defaultPlanFormProps } from '../../../../../../components/forms/PlanForm';
+import { generatePlanDefinition } from '../../../../../../components/forms/PlanForm/helpers';
 import HeaderBreadcrumb from '../../../../../../components/page/HeaderBreadcrumb/HeaderBreadcrumb';
-import { defaultProps as defaultPlanFormProps } from '../../../../../../containers/forms/ConnectedPlanForm';
 import store from '../../../../../../store';
-import BaseNewPlan, { NewIRSPlan, NewPlanForPlanning } from '../index';
+import {
+  addPlanDefinition,
+  removePlanDefinitions,
+} from '../../../../../../store/ducks/opensrp/PlanDefinition';
+import ConnectedBaseNewPlan, { NewIRSPlan, NewPlanForPlanning } from '../index';
+
+/* tslint:disable-next-line no-var-requires */
+const fetch = require('jest-fetch-mock');
 
 const history = createBrowserHistory();
+const middlewares: any = [];
+const mockStore = configureMockStore(middlewares);
 
 describe('containers/pages/NewPlan', () => {
   beforeEach(() => {
     jest.resetAllMocks();
+    fetch.resetMocks();
+    jest.clearAllMocks();
+    store.dispatch(removePlanDefinitions());
   });
 
   it('renders without crashing', () => {
-    shallow(<BaseNewPlan />);
+    shallow(<ConnectedBaseNewPlan />);
   });
 
   it('renders correctly', () => {
     const wrapper = mount(
       <Provider store={store}>
         <Router history={history}>
-          <BaseNewPlan />
+          <ConnectedBaseNewPlan />
         </Router>
       </Provider>
     );
@@ -32,7 +46,7 @@ describe('containers/pages/NewPlan', () => {
     // check that page title is displayed
     expect(toJson(wrapper.find('h3.mb-3.page-title'))).toMatchSnapshot('page title');
 
-    expect(wrapper.find('ConnectedPlanForm').props()).toEqual({
+    expect(wrapper.find('PlanForm').props()).toEqual({
       ...defaultPlanFormProps,
       addPlan: expect.any(Function),
       formHandler: expect.any(Function),
@@ -45,7 +59,7 @@ describe('containers/pages/NewPlan', () => {
       wrapper
         .find('Row')
         .find('Col#planform-col-container')
-        .find('ConnectedPlanForm')
+        .find('PlanForm')
     ).toHaveLength(1);
 
     // test that JurisdictionDetails are shown
@@ -84,7 +98,7 @@ describe('containers/pages/NewPlan', () => {
       </Provider>
     );
 
-    expect(wrapper.find('ConnectedPlanForm').props()).toEqual({
+    expect(wrapper.find('PlanForm').props()).toEqual({
       ...defaultPlanFormProps,
       addPlan: expect.any(Function),
       allowMoreJurisdictions: false,
@@ -112,7 +126,7 @@ describe('containers/pages/NewPlan', () => {
       </Provider>
     );
 
-    expect(wrapper.find('ConnectedPlanForm').props()).toEqual({
+    expect(wrapper.find('PlanForm').props()).toEqual({
       ...defaultPlanFormProps,
       addPlan: expect.any(Function),
       allowMoreJurisdictions: false,
@@ -138,5 +152,115 @@ describe('containers/pages/NewPlan', () => {
     expect(wrapper.find('JurisdictionDetails').length).toEqual(0);
 
     wrapper.unmount();
+  });
+
+  it('New plan is added to store if API status is 200', async () => {
+    fetch.mockResponseOnce(JSON.stringify({}), { status: 201 });
+
+    const initialState = {};
+    const mockedStore = mockStore(initialState);
+    mockedStore.dispatch = jest.fn();
+
+    const wrapper = mount(
+      <Provider store={mockedStore}>
+        <Router history={history}>
+          <ConnectedBaseNewPlan />
+        </Router>
+      </Provider>
+    );
+    // Set FI for interventionType
+    wrapper
+      .find('select[name="interventionType"]')
+      .simulate('change', { target: { name: 'interventionType', value: 'FI' } });
+    // set jurisdiction id ==> we use Formik coz React-Select is acting weird
+    (wrapper
+      .find('FieldInner')
+      .first()
+      .props() as any).formik.setFieldValue('jurisdictions[0].id', '1337');
+    // set jurisdiction name
+    wrapper
+      .find('input[name="jurisdictions[0].name"]')
+      .simulate('change', { target: { name: 'jurisdictions[0].name', value: 'Onyx' } });
+    // Set fiReason field value
+    wrapper
+      .find('select[name="fiReason"]')
+      .simulate('change', { target: { name: 'fiReason', value: 'Routine' } });
+    // Set fiStatus field value
+    wrapper
+      .find('select[name="fiStatus"]')
+      .simulate('change', { target: { name: 'fiStatus', value: 'A2' } });
+
+    wrapper.find('form').simulate('submit');
+
+    await new Promise<any>(resolve => setImmediate(resolve));
+
+    // no errors are initially shown
+    expect(
+      (wrapper
+        .find('FieldInner')
+        .first()
+        .props() as any).formik.errors
+    ).toEqual({});
+
+    // the expected payload
+    const payload = generatePlanDefinition(
+      (wrapper
+        .find('FieldInner')
+        .first()
+        .props() as any).formik.values
+    );
+
+    expect(mockedStore.dispatch).toHaveBeenLastCalledWith(addPlanDefinition(payload));
+  });
+
+  it('New plan is NOT added to store if API status is NOT 200', async () => {
+    fetch.mockReject(() => Promise.reject('API is down'));
+
+    const initialState = {};
+    const mockedStore = mockStore(initialState);
+    mockedStore.dispatch = jest.fn();
+
+    const wrapper = mount(
+      <Provider store={mockedStore}>
+        <Router history={history}>
+          <ConnectedBaseNewPlan />
+        </Router>
+      </Provider>
+    );
+    // Set FI for interventionType
+    wrapper
+      .find('select[name="interventionType"]')
+      .simulate('change', { target: { name: 'interventionType', value: 'FI' } });
+    // set jurisdiction id ==> we use Formik coz React-Select is acting weird
+    (wrapper
+      .find('FieldInner')
+      .first()
+      .props() as any).formik.setFieldValue('jurisdictions[0].id', '1337');
+    // set jurisdiction name
+    wrapper
+      .find('input[name="jurisdictions[0].name"]')
+      .simulate('change', { target: { name: 'jurisdictions[0].name', value: 'Onyx' } });
+    // Set fiReason field value
+    wrapper
+      .find('select[name="fiReason"]')
+      .simulate('change', { target: { name: 'fiReason', value: 'Routine' } });
+    // Set fiStatus field value
+    wrapper
+      .find('select[name="fiStatus"]')
+      .simulate('change', { target: { name: 'fiStatus', value: 'A2' } });
+
+    wrapper.find('form').simulate('submit');
+
+    await new Promise<any>(resolve => setImmediate(resolve));
+
+    // no errors are initially shown
+    expect(
+      (wrapper
+        .find('FieldInner')
+        .first()
+        .props() as any).formik.errors
+    ).toEqual({});
+
+    expect(mockedStore.dispatch).not.toHaveBeenCalled();
   });
 });
